Initial commit
Added the ability to play chess against stockfish, swap black v. white, and grab a screen capture and show what was grabbed.  The grabs are saved 1x per second.  The idea is to add the ability to recognize the pieces and use the latest image to reseed the stockfish board.